users-service/requirements.txt
asgiref==3.11.1
cryptography==45.0.7
Django==6.0.3
djangorestframework==3.16.1
PyJWT==2.10.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
